% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDmatrix.R
\name{getEDmatrix}
\alias{getEDmatrix}
\title{Calculate the Euclidean distances between two datasets}
\usage{
getEDmatrix(set1, set2)
}
\arguments{
\item{set1}{a data frame containing only the molecular features meant for the calculation of the Euclidean Distance}

\item{set2}{a data frame containing only the molecular features meant for the calculation of the Euclidean Distance}
}
\value{
a \code{getEDmatrix} object which consists of \code{set1} vs \code{set2} data frame of Euclidean distances, where the values in each row are sorted in
ascending order. As a consequence columns have no meaning on their own. \code{getEDmatrix} also implicitly creates two
variables, \code{maxs} and \code{mins}, which are automatically saved under such names and do not need explicitly variable
assignment. They are created for later data scaling.
}
\description{
Function that calculates a matrix of Euclidean distances between each pair of instances from two datasets.
}
\details{
No \code{NA} values are accepted, so either the respective instance is previously removed or empty values should be replaced
(eg., with the respective column median or average). For the purpose of using this package, \code{set1} and \code{set2} should be the same dataset.
All columns present at the data frames are used in the calculation of the Euclidean distances, i.e. Euclidean distance between set1[rowi] and set2[rowj].
Prior to calculating the Euclidean distances, the datasets will be scaled using \code{scale}, which applies \eqn{\frac{x_{ij}-min_j}{max_j-min_j}}
to each instance \eqn{x_i} under column (feature) j.
}
\examples{
train <- matrix(1:9,nrow=3, ncol=3)
a <- getEDmatrix(train, train)

}

